# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Devian Hub API",
    "description": "\nAPI do hub do **Devian** — projetos, chats (Claude Code) e artefatos de build.\n\n**Autenticação:** os endpoints exigem um **Bearer token** com dois formatos:\n\n1. **Token de máquina** (`DEVIAN_API_TOKEN`) — automação/scripts; atua como o admin.\n2. **JWT de sessão** — retornado por `POST /auth/login` após login com a conta\n   Google (via `google_sign_in` no app; sem Firebase). O backend valida o ID\n   token contra as chaves públicas do Google e emite o token de sessão.\n\nOs dados (projetos, chats, mensagens e artefatos) são escopados pelo usuário\nautenticado: cada usuário vê apenas os próprios dados, resolvidos\nautomaticamente a partir do token.\n\nEndpoints públicos: `GET /health` e `POST /auth/login`.\nClique em **Authorize** e cole o token, sem o prefixo `Bearer `.\n",
    "version": "0.6.0"
  },
  "servers": [
    {
      "url": "https://api.agapech.com.br/devian",
      "description": "Produção — via túnel Cloudflare (oracle-hermi)"
    },
    {
      "url": "/devian",
      "description": "Local — backend direto na porta 8088"
    }
  ],
  "paths": {
...
    "/projects/{project_id}/chats/{chat_id}/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Chat history (paginated)",
        "operationId": "list_messages_projects__project_id__chats__chat_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "description": "UUID da mensagem mais antiga da página atual. Passe o `next_cursor` retornado para buscar as anteriores.",
              "title": "Cursor",
              "example": "0191f3b2-4c3a-7b00-8000-000000000013",
              "type": "string",
              "format": "uuid",
              "nullable": true
            },
            "description": "UUID da mensagem mais antiga da página atual. Passe o `next_cursor` retornado para buscar as anteriores."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Quantidade de mensagens por página (1–200)",
              "default": 50,
              "title": "Limit",
              "example": 50
            },
            "description": "Quantidade de mensagens por página (1–200)"
          }
        ],
        "responses": {
          "200": {
            "description": "Página de mensagens em ordem cronológica. Sem `cursor` → as **mais recentes**; com `cursor` (UUID da msg mais antiga da página) → as anteriores (scroll p/ cima). `next_cursor: null` = início da conversa.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou chat não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send message",
        "operationId": "send_message_projects__project_id__chats__chat_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mensagem do usuário gravada, processada pelo Claude Code na sessão do chat e resposta do assistente gravada e retornada. Payload leve: **só a última mensagem** — o contexto fica na sessão + no banco.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou chat não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Corpo da requisição inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "502": {
            "description": "Erro do Claude Code no container (ClaudeError)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "504": {
            "description": "O Claude Code não respondeu a tempo (timeout)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem de erro (em português)"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorOut",
        "description": "Resposta de erro padrão da API.",
        "example": {
          "message": "Projeto não encontrado"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "Message text sent to the assistant."
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MessageCreate",
        "description": "The user's next message. That's it — no history in the payload.",
        "example": {
          "content": "Pode adicionar um botão de exportar PDF no relatório?"
        }
      },
      "MessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "'user' or 'assistant'"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "title": "MessageOut",
        "example": {
          "content": "Claro! Adicionei um botão de exportar PDF no rodapé do relatório.",
          "created_at": "2026-08-05T16:00:05Z",
          "id": "0191f3b2-4c3a-7b00-8000-000000000014",
          "role": "assistant"
        }
      },
      "MessagePage": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Messages"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "messages",
          "next_cursor"
        ],
        "title": "MessagePage",
        "description": "History page. `next_cursor` = id (UUID) of the oldest message in this\npage; pass it as `cursor` on the next call to fetch older ones\n(scroll up). `null` = beginning of the conversation reached.",
        "example": {
          "messages": [
            {
              "content": "Claro! Feito.",
              "created_at": "2026-08-05T16:00:05Z",
              "id": "0191f3b2-4c3a-7b00-8000-000000000014",
              "role": "assistant"
            },
            {
              "content": "Pode adicionar um botão de exportar PDF?",
              "created_at": "2026-08-05T16:00:04Z",
              "id": "0191f3b2-4c3a-7b00-8000-000000000013",
              "role": "user"
            }
          ],
          "next_cursor": "0191f3b2-4c3a-7b00-8000-000000000013"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Token Bearer do Devian — aceita dois formatos:\n1. Token de máquina (DEVIAN_API_TOKEN): atua como o admin (dono).\n2. JWT de sessão: retornado por POST /auth/login após login com Google.\nCole apenas o token, sem o prefixo 'Bearer '.",
        "scheme": "bearer"
      }
    }
  },
...
}